requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.22
colorama==0.4.6
duckdb==0.7.1
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lz4==4.3.2
MarkupSafe==2.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
openai==0.27.5
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
posthog==3.0.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
zstandard==0.21.0
